refactor(web): extract useTraining hook from App.jsx (sc-1651)
Third Phase-A slice. Moves the project's training-dataset state
(trainingDatasets, trainingDatasetsProjectId, loadingTrainingDatasets,
trainingDatasetsError) and the 8 dataset/job functions (refreshTrainingDatasets,
loadTrainingDataset, create/update/batchRename/writeCaptionSidecars dataset,
createTrainingDatasetCaptionJob, createTrainingJob) into hooks/useTraining.js.
App composes useTraining({ token, activeProject, setError, setJobs }) and forwards
the values down unchanged, so TrainingStudio is untouched. The training target/
preset catalogs (bulk-loaded by refreshData) stay in App; caption/training jobs
still push onto the shared jobs list via the injected setJobs.

Behavior-preserving: App.jsx 2132 -> 2026 lines (god module now 2334 -> 2026,
-308 across three slices). 107 web vitest pass, build clean, browser smoke (Rust
API on a temp data dir): Training Studio renders with all hook props wired,
refreshTrainingDatasets loads a seeded dataset on project load (0 -> 1 dataset)
with no console errors.
